requirements .txt
# Core Scraping
Scrapy>=2.5.0,<3.0.0
itemloaders>=1.0.0,<2.0.0
crochet>=2.1.0,<3.0.0

# API
fastapi>=0.68.0,<0.69.0
uvicorn[standard]>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
python-multipart>=0.0.5,<0.0.6

# Database
psycopg2-binary>=2.9.0,<3.0.0
sqlalchemy>=1.4.0,<2.0.0

# Scheduler
apscheduler>=3.8.0,<4.0.0

# Utils
python-dotenv>=0.19.0,<0.20.0
requests>=2.26.0,<3.0.0
beautifulsoup4>=4.9.3,<5.0.0
lxml>=4.6.3,<5.0.0

# Development
pytest>=6.2.5,<7.0.0
pytest-cov>=2.12.1,<3.0.0
black>=21.7b0,<22.0.0
flake8>=3.9.2,<4.0.0
